requirements_prod.txt
Django==4.2.2
django-bootstrap5==23.3
django-cachalot==2.6.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-json-widget==1.1.1
django-redis==5.3.0
celery==5.3.1
flower==2.0.0
pandas==2.0.2
pdfkit==1.0.0
Pillow==9.5.0
psycopg2==2.9.6
python-dotenv==1.0.0
requests==2.31.0
redis==4.6.0
easy-thumbnails==2.8.5
openpyxl==3.1.2
passlib==1.7.4

gunicorn==21.2.0

djangorestframework==3.14.0
django-filter==23.2

social-auth-app-django==5.2.0

#exchangers
pybit==5.3.0
python-binance==1.0.17
python-okx==0.2.0
mexcapiclient==0.1.9

cryptography==41.0.1
pycryptodome==3.18.0